move chat to it's own page
