+++ client/src/App.tsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect, useContext } from 'react';
+import InviteToChat from './components/InviteToChat';
 import HomePage from './components/Home Page/HomePage';
 import PulloutMenu from './components/Home Page/PulloutMenu';
 import RSSFeed from './components/rss/RSSFeedContainer';
@@ -13,32 +14,23 @@ import ThemeProvider from '@mui/material/styles/ThemeProvider';
 import createTheme from '@mui/material/styles/createTheme';
 import { PaletteOptions } from '@mui/material';
 import { light, dark, veggie, meat } from './Theme';
-import io from 'socket.io-client';
-import Chat from './components/Chat';
-import './App.css';
+
 import ClimbingBoxLoader from 'react-spinners/ClimbingBoxLoader';
 
-const socket = io.connect('http://localhost:3001', { path: '/socket.io' });
 interface ThemeOptions {
   palette?: PaletteOptions;
 }
 
 const App: React.FC = (): JSX.Element => {
   const { getUser, user, userAccount, loggedIn } = useContext(UserContext);
-  const [username, setUsername] = useState('');
-  const [room, setRoom] = useState('');
-  const [showChat, setShowChat] = useState(false);
-  // socket is what we call the actual connection to the socket server
 
   const [loading, setLoading] = useState(false);
   const [color, setColor] = useState('#1682B2');
-  // const { userAccount, loggedIn } = useContext(UserContext);
+  const { userAccount, loggedIn } = useContext(UserContext);
 
   useEffect(() => {
     userAccount();
   }, [loggedIn]);
-
-  // console.log(user, 'app tsx 41');
 
   useEffect(() => {
     setLoading(true);
@@ -50,13 +42,6 @@ const App: React.FC = (): JSX.Element => {
   const [theme, setTheme] = useState<ThemeOptions>(light);
 
   const chosenTheme = createTheme(theme);
-
-  const joinRoom = () => {
-    if (username !== '' && room !== '') {
-      socket.emit('join_room', room);
-      setShowChat(true);
-    }
-  };
 
   return (
     <ThemeProvider theme={chosenTheme}>
@@ -76,30 +61,6 @@ const App: React.FC = (): JSX.Element => {
       ) : (
         <div>
           <PulloutMenu changeTheme={setTheme} />
-          <div className='chatContainer'>
-            {!showChat ? (
-              <div className='joinChatContainer'>
-                <h3>Live Chat</h3>
-                <input
-                  type='text'
-                  placeholder='Room ID...'
-                  onChange={(e) => {
-                    setRoom(e.target.value);
-                  }}
-                />
-                <input
-                  type='text'
-                  placeholder='Username'
-                  onChange={(e) => {
-                    setUsername(e.target.value);
-                  }}
-                />
-                <button onClick={joinRoom}>Join A Room</button>
-              </div>
-            ) : (
-              <Chat socket={socket} username={username} room={room} />
-            )}
-          </div>
           <Switch>
             <Route exact path='/'>
               <HomePage />
@@ -121,6 +82,9 @@ const App: React.FC = (): JSX.Element => {
             </Route>
             <Route path='/market_finder'>
               <Map />
+            </Route>
+            <Route path='/live_chat'>
+              <InviteToChat />
             </Route>
           </Switch>
         </div>
